requirements.txt
fire==0.1.2
PyYAML==3.12
Scrapy==1.4.0
